+++ dropbot_status/dock_pane.py
@@ -15,13 +15,13 @@ class DropbotStatusDockPane(DockPane):
 
     def create_contents(self, parent):
         # Import all the components from your module
-        from .widget import (
-            DropBotStatusModel,
+        from .displayed_UI import (
             DropbotStatusViewModelSignals,
             DropBotStatusViewModel,
             DropBotStatusView
         )
-        from .dramatiq_widget import DramatiqDropBotStatusViewModel, DramatiqDropBotStatusView, DramatiqDropBotStatusViewModelSignals
+        from .model import DropBotStatusModel
+        from .dramatiq_UI import DramatiqDropBotStatusViewModel, DramatiqDropBotStatusView, DramatiqDropBotStatusViewModelSignals
         from .dramatiq_dropbot_status_controller import DramatiqDropbotStatusController
 
         model = DropBotStatusModel()
@@ -32,6 +32,8 @@ class DropbotStatusDockPane(DockPane):
             model=model,
             view_signals=dramatiq_view_signals
         )
+
+        # store controller and view in dock pane
         self.dramatiq_controller = DramatiqDropbotStatusController(ui=dramatiq_view_model,
                                                                    listener_name=dramatiq_view_model.__class__.__module__.split(".")[0] + "_listener")
         self.dramatiq_status_view = DramatiqDropBotStatusView(view_model=dramatiq_view_model)
